requirements.txt
PyMySQL~=1.0.3
requests~=2.31.0
python-dateutil~=2.8.2